run elm app separately
mix of elm app and bruch was not efficient enough and it took long time
to recompile the project
