+++ assets/brunch-config.js
@@ -38,7 +38,7 @@ exports.config = {
   // Phoenix paths configuration
   paths: {
     // Dependencies and current project directories to watch
-    watched: ["static", "css", "js", "vendor", "elm"],
+    watched: ["static", "css", "js", "vendor"],
     // Where to compile files to
     public: "../priv/static"
   },
@@ -49,12 +49,6 @@ exports.config = {
       // Do not use ES6 compiler in vendor code
       ignore: [/vendor/]
     },
-    elmBrunch: {
-      executablePath: "./node_modules/.bin",
-      mainModules: ["elm/Main.elm"],
-      outputFolder: "js",
-      makeParameters: ['--debug'],
-    }
   },
 
   modules: {
