+++ assets/elm-app/src/User/LoginPage.elm
@@ -7,7 +7,6 @@ import Html.Attributes exposing (..)
 import Html.Events exposing (onInput, onSubmit)
 import Json.Decode as Decode exposing (Decoder)
 import Json.Decode.Pipeline as Pipeline exposing (decode, optional)
-
 import Helpers.Views.Form as Form
 import User.SessionData exposing (Session)
 import User.UserData exposing (User)
@@ -17,10 +16,22 @@ import Helpers.Request.ErrorsData as ErrorsData
 import Route
 
 
+-- Material
+
+import Material.Grid as Grid exposing (align, offset, grid, cell, Device(..))
+import Material.Options as Options exposing (cs, css, Style, when)
+import Material.Elevation as Elevation
+import Material.Button as Button
+import Material.Typography as Typo
+import Material
+import Material.Textfield as Textfield
+
+
 type alias Model =
     { errors : List Error
     , username : String
     , password : String
+    , mdl : Material.Model
     }
 
 
@@ -29,6 +40,7 @@ initialModel =
     { errors = []
     , username = ""
     , password = ""
+    , mdl = Material.model
     }
 
 
@@ -39,7 +51,8 @@ type Field
 
 
 type Msg
-    = SetUsername String
+    = Mdl (Material.Msg Msg)
+    | SetUsername String
     | SetPassword String
     | SubmitForm
     | LoginCompleted (Result Http.Error User)
@@ -60,6 +73,67 @@ type ExternalMsg
 
 view : Session -> Model -> Html Msg
 view session model =
+    div []
+        [ grid []
+            [ cell [ Grid.size Phone 4, Grid.size Tablet 6, offset Tablet 1, Grid.size Desktop 8, offset Desktop 2 ]
+                [ h4 [] [ text "Sign in to GitWerk" ]
+                ]
+            ]
+        , Options.div
+            [ Elevation.e4
+            , Options.center
+            , css "width" "408px"
+            , css "margin" "auto"
+            ]
+            [ grid
+                []
+                [ cell [ Grid.size All 12 ]
+                    [ div
+                        []
+                        [ Textfield.render Mdl
+                            [ 0 ]
+                            model.mdl
+                            [ Textfield.label "username"
+                            , Textfield.floatingLabel
+                            , Textfield.text_
+                            ]
+                            []
+                        ]
+                    , div []
+                        [ Textfield.render Mdl
+                            [ 2 ]
+                            model.mdl
+                            [ Textfield.label "Enter password"
+                            , Textfield.floatingLabel
+                            , Textfield.password
+                            , Options.onInput SetPassword
+                            ]
+                            []
+                        ]
+                    , div []
+                        [ Button.render Mdl
+                            [ 3 ]
+                            model.mdl
+                            [ Button.raised
+                            , Button.ripple
+                            , Button.colored
+                            , Options.onClick SubmitForm
+                            ]
+                            [ text "Sign in" ]
+                        ]
+                    , Options.styled div
+                        [ Typo.body1 ]
+                        [ text "or "
+                        , a [ Route.href Route.Join ] [ text "create an account" ]
+                        ]
+                    ]
+                ]
+            ]
+        ]
+
+
+view2 : Session -> Model -> Html Msg
+view2 session model =
     div [ class "auth-page" ]
         [ div [ class "container page" ]
             [ div [ class "row" ]
@@ -117,7 +191,8 @@ update msg model =
 
         LoginCompleted (Err error) ->
             let
-                errorMessages = ErrorsData.httpErrorToList "registration" error errorsDecoder
+                errorMessages =
+                    ErrorsData.httpErrorToList "registration" error errorsDecoder
             in
                 { model | errors = List.map (\errorMessage -> Form => errorMessage) errorMessages }
                     => Cmd.none
@@ -125,12 +200,16 @@ update msg model =
 
         LoginCompleted (Ok user) ->
             model
-            => Cmd.batch [UserRequest.storeSession user, Route.modifyUrl Route.Home]
-            => SetUser user
+                => Cmd.batch [ UserRequest.storeSession user, Route.modifyUrl Route.Home ]
+                => SetUser user
+
+        Mdl msg_ ->
+            Material.update Mdl msg_ model
+                => NoOp
+
 
 errorsDecoder : Decoder (List String)
 errorsDecoder =
     decode (\username password -> List.concat [ username, password ])
         |> ErrorsData.optionalError "username"
         |> ErrorsData.optionalError "password"
-
