+++ assets/elm-app/tests/Tests.elm
@@ -1,0 +1,23 @@
+module Tests exposing (..)
+
+import Test exposing (..)
+import Expect
+import String
+
+
+-- Check out http://package.elm-lang.org/packages/elm-community/elm-test/latest to learn more about testing in Elm!
+
+
+all : Test
+all =
+    describe "A Test Suite"
+        [ test "Addition" <|
+            \_ ->
+                Expect.equal 10 (3 + 7)
+        , test "String.left" <|
+            \_ ->
+                Expect.equal "a" (String.left 1 "abcdefg")
+        , test "This test should fail" <|
+            \_ ->
+                Expect.fail "failed as expected!"
+        ]
