+++ assets/elm-app/tests/elm-package.json
@@ -4,8 +4,8 @@
     "repository": "https://github.com/user/project.git",
     "license": "BSD3",
     "source-directories": [
-        "../elm",
-        "."
+        ".",
+        "../src"
     ],
     "exposed-modules": [],
     "dependencies": {
@@ -16,7 +16,8 @@
         "evancz/url-parser": "2.0.1 <= v < 3.0.0",
         "NoRedInk/elm-decode-pipeline": "3.0.0 <= v < 4.0.0",
         "elm-lang/http": "1.0.0 <= v < 2.0.0",
-        "lukewestby/elm-http-builder": "5.1.0 <= v < 6.0.0"
+        "lukewestby/elm-http-builder": "5.1.0 <= v < 6.0.0",
+        "debois/elm-mdl": "8.1.0 <= v < 9.0.0"
     },
     "elm-version": "0.18.0 <= v < 0.19.0"
 }
